hansardparser/plenaryparser/requirements.txt
chardet==3.0.4
Flask==1.0.2
google-api-python-client==1.7.8
google-cloud-storage==1.14.0
gunicorn==19.9.0
nltk==3.2.1
numpy==1.16.2
pandas==0.24.2
PyPDF2==1.26.0
requests==2.21.0
tensor2tensor==1.13.2
tensorboard==1.13.1
tensorflow==1.13.1
tensorflow-datasets==1.0.1
tensorflow-estimator==1.13.0
tensorflow-hub==0.4.0
tensorflow-metadata==0.13.0
tensorflow-probability==0.6.0
tqdm==4.31.1
